Remove rocket landing strat, keep only scaffold
The proposed "strategy" is to choose a landing location based on its
degree. The reasoning is that it would be pointless to land in a
location where it would be impossible or difficult to unload units
efficiently.

However, this is currently removed (all locations are simply assigned
the highest degree), as in some rare cases, you may wish to dump
rockets in bad locations, due simply to the fact that they would be
inaccessible - no enemy units come close enough to destroy them. This
would contribute to the overall value.

How the strategy would be implemented is up for discussion.